Add selecting to state educational reqs
Moves away from hard-coded info. Now the state info is stored in the DB and is
editable by the user. The stupid thing I did here was not making these routable.
By doing it all in JS I made it impossible to link to a single state's educational
reqs. Work for the future.